feat(sandbox): out-of-box UX — fallback warn, bundled bwrap, doctor CLI, helper CI, readiness guides, WSL2 routing
Batch A: default fallback deny→warn. Commands run with warning instead of hard error when sandbox unavailable.
Batch B: bundled bwrap discovery in Rust helper. CI build+hash pipeline for helpers. Placeholder CI directory.
Batch C: synergy doctor CLI with friendly readiness output. Readiness detail includes exact install commands for missing deps.
Batch D: WSL2 detection routes sandbox to Linux bwrap backend. Readiness reflects WSL2 routing.
