+++ packages/synergy/package.json
@@ -15,6 +15,7 @@
     "format": "echo 'Formatting code...' && bun run --prettier --write src/**/*.ts",
     "docs": "echo 'Generating documentation...' && find src -name '*.ts' -exec echo 'Processing: {}' \\;",
     "hash-helper": "bun run scripts/hash-helper.ts",
+    "build-helper": "bun run scripts/build-helper.ts",
     "deploy": "echo 'Deploying application...' && bun run build && echo 'Deployment completed successfully'"
   },
   "bin": {
