+++ packages/synergy/scripts/hash-helper.ts
@@ -1,5 +1,6 @@
 #!/usr/bin/env bun
 import * as fs from "fs"
+import * as path from "path"
 import { createHash } from "crypto"
 
 // Usage:
@@ -27,6 +28,18 @@ const helperName = platform === "linux" ? "synergy-sandbox-linux" : "synergy-san
 console.log(`export const ${constName}: Record<string, string> = {`)
 console.log(`  path.join(homedir, ".synergy", "sandbox-helper", "${helperName}"): "${digest}",`)
 console.log(`}`)
+
+if (process.argv.includes("--generate")) {
+  const sourceFile = path.resolve(import.meta.dir, "..", "src", "sandbox", targetModule)
+  let content = fs.readFileSync(sourceFile, "utf-8")
+  const pattern = new RegExp(`export const ${constName}: Record<string, string> = \\{[^}]*\\}`, "s")
+  const replacement = `export const ${constName}: Record<string, string> = {
+  path.join(homedir, ".synergy", "sandbox-helper", "${helperName}"): "${digest}",
+}`
+  content = content.replace(pattern, replacement)
+  fs.writeFileSync(sourceFile, content)
+  console.log(`Updated ${sourceFile} with new hash entry`)
+}
 
 // Also output the JSON format for potential future release automation
 console.log()
