+++ packages/synergy/src/sandbox/backend.ts
@@ -58,6 +58,9 @@ import { MacBackend } from "./macos"
 import { LinuxBackend } from "./linux"
 import { WindowsBackend } from "./windows"
 import { startDenialLogger, type DenialLoggerSession } from "./macos-diagnostics"
+import { Log } from "@/util/log"
+import { isWsl2 } from "./wsl"
+const log = Log.create({ service: "sandbox-backend" })
 
 // ------------------------------------------------------------------
 // SandboxBackend — unified public API
@@ -96,7 +99,14 @@ export namespace SandboxBackend {
       return { command: opts.command, args: opts.args, sandboxed: false }
     }
 
-    const platform = opts.forcePlatform ?? detectPlatform()
+    let platform = opts.forcePlatform ?? detectPlatform()
+
+    // WSL2 → route to Linux sandbox backend
+    if (platform === "windows" && isWsl2()) {
+      platform = "linux"
+      log.info("sandbox.wsl2_routing", { original: "windows", routed: "linux" })
+    }
+
     switch (platform) {
       case "macos":
         return MacBackend.prepare(opts)
