+++ packages/synergy/src/sandbox/helper-linux/.gitignore
@@ -1,0 +1,1 @@
+bwrap/bwrap
