+++ packages/synergy/src/server/sandbox-readiness-route.ts
@@ -1,14 +1,8 @@
 import { Hono } from "hono"
 import { describeRoute, resolver } from "hono-openapi"
 import z from "zod"
-import * as fs from "fs"
 import { Config } from "../config/config"
-import { detectPlatform } from "../sandbox/detect"
-import { platformInfo, isBwrapAvailable } from "../sandbox/platform"
-import { getWindowsHelperInfo } from "../sandbox/windows"
-import { getLinuxHelperInfo, TRUSTED_LINUX_HELPER_HASHES } from "../sandbox/linux"
-import { isWsl1 } from "../sandbox/wsl"
-import type { SandboxReadinessCheck } from "../sandbox/types"
+import { getSandboxReadiness } from "../sandbox/readiness"
 
 const SandboxReadinessCheckSchema = z
   .object({
@@ -16,6 +10,13 @@ const SandboxReadinessCheckSchema = z
     label: z.string(),
     status: z.enum(["pass", "warn", "fail"]),
     detail: z.string(),
+    recovery: z
+      .object({
+        action: z.string(),
+        label: z.string(),
+        command: z.string(),
+      })
+      .optional(),
   })
   .meta({ ref: "SandboxReadinessCheck" })
 
@@ -48,377 +49,8 @@ export const SandboxReadinessRoute = new Hono().get(
     },
   }),
   async (c) => {
-    const platform = detectPlatform()
-    const info = platformInfo()
     const cfg = await Config.get()
-    const sandboxCfg = cfg.sandbox
-    const checks: SandboxReadinessCheck[] = []
-
-    // Normalize platform name for the typed response
-    let platformName: "macos" | "linux" | "windows" | "unsupported"
-    if (platform === "macos" || platform === "linux" || platform === "windows") {
-      platformName = platform
-    } else {
-      platformName = "unsupported"
-    }
-
-    // ── Generic: sandbox config check ───────────────────────────
-    const configEnabled = sandboxCfg?.enabled !== false // defaults to true when absent
-    if (!sandboxCfg) {
-      checks.push({
-        id: "sandbox_config",
-        label: "Sandbox configuration",
-        status: "warn",
-        detail: "No sandbox configuration section found. Using platform defaults (enabled=true).",
-      })
-    } else if (!configEnabled) {
-      checks.push({
-        id: "sandbox_config",
-        label: "Sandbox configuration",
-        status: "fail",
-        detail: "Sandbox is disabled via configuration (enabled: false).",
-      })
-    } else {
-      checks.push({
-        id: "sandbox_config",
-        label: "Sandbox configuration",
-        status: "pass",
-        detail: `Sandbox enabled with backend "${sandboxCfg.backend ?? "auto"}", fallback: ${sandboxCfg.fallbackPolicy ?? "deny"}.`,
-      })
-    }
-
-    // ── Platform-specific checks ────────────────────────────────
-    switch (platform) {
-      case "macos": {
-        // Check sandbox-exec binary
-        const sandboxExecExists = (() => {
-          try {
-            return fs.existsSync("/usr/bin/sandbox-exec")
-          } catch {
-            return false
-          }
-        })()
-        checks.push({
-          id: "macos_sandbox_exec",
-          label: "sandbox-exec binary",
-          status: sandboxExecExists ? "pass" : "fail",
-          detail: sandboxExecExists
-            ? "/usr/bin/sandbox-exec found"
-            : "/usr/bin/sandbox-exec not found — macOS sandbox unavailable",
-        })
-
-        // Check sandbox-exec version / deprecation note
-        checks.push({
-          id: "macos_sandbox_exec_version",
-          label: "sandbox-exec version",
-          status: sandboxExecExists ? "warn" : "fail",
-          detail: sandboxExecExists
-            ? "sandbox-exec available (deprecated since macOS 10.12 Sierra but still functional on Apple Silicon)"
-            : "sandbox-exec not found",
-        })
-
-        // Check log stream utility for denial logging
-        const logExists = (() => {
-          try {
-            return fs.existsSync("/usr/bin/log")
-          } catch {
-            return false
-          }
-        })()
-        checks.push({
-          id: "macos_log_stream",
-          label: "Denial logger (log stream)",
-          status: logExists ? "pass" : "warn",
-          detail: logExists
-            ? "/usr/bin/log found — denial logging available"
-            : "/usr/bin/log not found — denial logging unavailable (non-critical)",
-        })
-        break
-      }
-      case "linux": {
-        // ── WSL1 detection ────────────────────────────────────
-        // WSL1 lacks the kernel features (namespaces, seccomp)
-        // required for bwrap-based sandboxing.
-        const wsl1Detected = isWsl1()
-        checks.push({
-          id: "linux_wsl1",
-          label: "WSL1 environment",
-          status: wsl1Detected ? "fail" : "pass",
-          detail: wsl1Detected
-            ? "WSL1 detected — bwrap namespace sandboxing is unsupported due to kernel limitations. Upgrade to WSL2 for full sandbox support."
-            : "Not running under WSL1",
-        })
-        // ── Linux helper binary check ──────────────────────────
-        const linuxHelperInfo = getLinuxHelperInfo()
-        if (!linuxHelperInfo) {
-          checks.push({
-            id: "linux_helper",
-            label: "Linux sandbox helper binary",
-            status: "fail",
-            detail: "synergy-sandbox-linux helper not found. Install the Synergy sandbox helper for Linux.",
-            recovery: {
-              type: "install_helper",
-              backend: "synergy-sandbox-linux",
-              instructions: "Install the Synergy sandbox helper for Linux",
-            },
-          })
-          checks.push({
-            id: "linux_helper_hash",
-            label: "Linux helper hash verification",
-            status: "fail",
-            detail: "Cannot verify hash — helper binary not found.",
-          })
-        } else if (!linuxHelperInfo.verified) {
-          checks.push({
-            id: "linux_helper",
-            label: "Linux sandbox helper binary",
-            status: "warn",
-            detail: `Helper binary found at ${linuxHelperInfo.path} but hash verification failed.`,
-            recovery: {
-              type: "install_helper",
-              backend: "synergy-sandbox-linux",
-              instructions:
-                "Sandbox helper binary hash verification failed. The helper may be corrupted or tampered. Reinstall the Synergy Linux sandbox helper.",
-            },
-          })
-          checks.push({
-            id: "linux_helper_hash",
-            label: "Linux helper hash verification",
-            status: "fail",
-            detail: "Hash verification failed — helper may be corrupted or tampered.",
-          })
-        } else {
-          checks.push({
-            id: "linux_helper",
-            label: "Linux sandbox helper binary",
-            status: "pass",
-            detail: `Helper binary found and verified at ${linuxHelperInfo.path}.`,
-          })
-          checks.push({
-            id: "linux_helper_hash",
-            label: "Linux helper hash verification",
-            status: "pass",
-            detail: "Helper hash verified against trusted hashes.",
-          })
-        }
-
-        // ── Linux seccomp check ─────────────────────────────────
-        const seccompAvailable = (() => {
-          try {
-            // Check if /proc/sys/kernel/seccomp/actions_avail exists
-            // This file lists available seccomp actions (e.g. kill_process, trap, errno, trace, log, allow)
-            const result = Bun.spawnSync({
-              cmd: ["cat", "/proc/sys/kernel/seccomp/actions_avail"],
-              stdout: "pipe",
-              stderr: "pipe",
-            })
-            if (result.exitCode === 0) {
-              const val = new TextDecoder().decode(result.stdout).trim()
-              // SECCOMP_SET_MODE_FILTER corresponds to actions "errno", "trap", "kill", "kill_process", "allow", "trace", "log"
-              return val.length > 0 && val.includes("errno")
-            }
-            // Fallback: check classic /proc/sys/kernel/seccomp/actions_logged (older kernels)
-            return false
-          } catch {
-            return false
-          }
-        })()
-        checks.push({
-          id: "linux_seccomp",
-          label: "seccomp BPF filtering",
-          status: seccompAvailable ? "pass" : "warn",
-          detail: seccompAvailable
-            ? "seccomp filter mode available"
-            : "seccomp filter mode unavailable — kernel may not support SECCOMP_SET_MODE_FILTER",
-        })
-
-        // ── Existing checks (bwrap, namespaces, landlock) ──────
-        // Check bwrap availability
-        const bwrapAvailable = isBwrapAvailable()
-        checks.push({
-          id: "linux_bwrap",
-          label: "bwrap (bubblewrap)",
-          status: bwrapAvailable ? "pass" : "fail",
-          detail: bwrapAvailable
-            ? "bwrap found on PATH"
-            : "bwrap not found — install bubblewrap for Linux sandbox support",
-        })
-
-        // Check user namespaces
-        const usernsAvailable = (() => {
-          try {
-            const result = Bun.spawnSync({
-              cmd: ["cat", "/proc/sys/kernel/unprivileged_userns_clone"],
-              stdout: "pipe",
-              stderr: "pipe",
-            })
-            if (result.exitCode === 0) {
-              const val = new TextDecoder().decode(result.stdout).trim()
-              return val === "1"
-            }
-            // File doesn't exist on newer kernels (>=4.x) where namespaces are always available
-            return true
-          } catch {
-            return true // If we can't check, assume enabled (most distros ship with it)
-          }
-        })()
-        checks.push({
-          id: "linux_namespaces",
-          label: "User namespaces",
-          status: usernsAvailable ? "pass" : "fail",
-          detail: usernsAvailable
-            ? "User namespaces appear enabled"
-            : "User namespaces disabled — bwrap requires unprivileged user namespaces. Set kernel.unprivileged_userns_clone=1.",
-        })
-
-        // Check max_user_namespaces (RHEL/Fedora)
-        const maxUsernsAvailable = (() => {
-          try {
-            const result = Bun.spawnSync({
-              cmd: ["cat", "/proc/sys/user/max_user_namespaces"],
-              stdout: "pipe",
-              stderr: "pipe",
-            })
-            if (result.exitCode === 0) {
-              const val = parseInt(new TextDecoder().decode(result.stdout).trim())
-              return val > 0
-            }
-            return true // file doesn't exist = unlimited (kernel default)
-          } catch {
-            return true
-          }
-        })()
-        checks.push({
-          id: "linux_user_max_namespaces",
-          label: "Max user namespaces",
-          status: maxUsernsAvailable ? "pass" : "fail",
-          detail: maxUsernsAvailable
-            ? "user.max_user_namespaces > 0"
-            : "user.max_user_namespaces = 0 — no user namespaces available. Set to > 0 or disable this restriction.",
-        })
-
-        // Check bwrap version
-        const bwrapVersion = (() => {
-          try {
-            const result = Bun.spawnSync({
-              cmd: ["bwrap", "--version"],
-              stdout: "pipe",
-              stderr: "pipe",
-            })
-            return new TextDecoder().decode(result.stdout).trim()
-          } catch {
-            return null
-          }
-        })()
-        checks.push({
-          id: "linux_bwrap_version",
-          label: "bwrap version",
-          status: bwrapVersion ? "pass" : "warn",
-          detail: bwrapVersion ?? "Could not determine bwrap version",
-        })
-
-        // Check Landlock LSM availability
-        const landlockAvailable = (() => {
-          try {
-            const result = Bun.spawnSync({
-              cmd: [
-                "sh",
-                "-c",
-                "grep -q landlock /proc/filesystems 2>/dev/null || grep -q CONFIG_SECURITY_LANDLOCK /boot/config-$(uname -r) 2>/dev/null",
-              ],
-              stdout: "pipe",
-              stderr: "pipe",
-            })
-            return result.exitCode === 0
-          } catch {
-            return false
-          }
-        })()
-        checks.push({
-          id: "linux_landlock",
-          label: "Landlock LSM",
-          status: landlockAvailable ? "pass" : "warn",
-          detail: landlockAvailable
-            ? "Landlock kernel module detected — available as fallback sandbox"
-            : "Landlock not detected — bwrap is the only Linux sandbox option",
-        })
-        break
-      }
-
-      case "windows": {
-        // Check helper binary existence and hash verification
-        const helperInfo = getWindowsHelperInfo()
-        if (!helperInfo) {
-          checks.push({
-            id: "windows_helper",
-            label: "Sandbox helper binary",
-            status: "fail",
-            detail: "synergy-sandbox-windows.exe not found. Install the Synergy Windows sandbox helper.",
-            recovery: {
-              type: "install_helper",
-              backend: "synergy-sandbox-windows",
-              instructions: "Install the Synergy sandbox helper for Windows",
-            },
-          })
-          checks.push({
-            id: "windows_helper_hash",
-            label: "Helper hash verification",
-            status: "fail",
-            detail: "Cannot verify hash — helper binary not found.",
-          })
-        } else if (!helperInfo.verified) {
-          checks.push({
-            id: "windows_helper",
-            label: "Sandbox helper binary",
-            status: "warn",
-            detail: `Helper binary found at ${helperInfo.path} but hash verification failed.`,
-            recovery: {
-              type: "install_helper",
-              backend: "synergy-sandbox-windows",
-              instructions:
-                "Sandbox helper binary hash verification failed. The helper may be corrupted or tampered. Reinstall the Synergy Windows sandbox helper.",
-            },
-          })
-          checks.push({
-            id: "windows_helper_hash",
-            label: "Helper hash verification",
-            status: "fail",
-            detail: "Hash verification failed — helper may be corrupted or tampered.",
-          })
-        } else {
-          checks.push({
-            id: "windows_helper",
-            label: "Sandbox helper binary",
-            status: "pass",
-            detail: `Helper binary found and verified at ${helperInfo.path}.`,
-          })
-          checks.push({
-            id: "windows_helper_hash",
-            label: "Helper hash verification",
-            status: "pass",
-            detail: "Helper hash verified against trusted hashes.",
-          })
-        }
-        break
-      }
-    }
-
-    // Ready: config must be enabled and no checks must have failed
-    const ready = configEnabled && !checks.some((c) => c.status === "fail")
-    // Compute summary
-    const summary = ready
-      ? "Sandbox is operational"
-      : checks
-          .filter((c) => c.status === "fail")
-          .map((c) => c.label)
-          .join(", ") + " — sandbox not ready"
-
-    return c.json({
-      platform: platformName,
-      backend: info.backend,
-      ready,
-      summary,
-      checks,
-    })
+    const readiness = await getSandboxReadiness(cfg.sandbox)
+    return c.json(readiness)
   },
 )
